+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixes
 - Some minor style fixes
 - Fixed the input field overflow in the project item badge editor
+- New e2e tests for inline editing
 
 0.1.4 - 2024-11-01
 ==================